+++ Code/SimpleAuthentication.Core/Providers/Google/AccessTokenResult.cs
@@ -13,7 +13,7 @@
 
         public string AccessToken { get; set; }
         public string TokenType { get; set; }
-        public int ExpiresIn { get; set; }
+        public long ExpiresIn { get; set; }
         public string IdToken { get; set; }
     }
 }
