+++ Code/SimpleAuthentication.ExtraProviders/LinkedIn/AccessTokenResult.cs
@@ -3,6 +3,6 @@
     public class AccessTokenResult
     {
         public string AccessToken { get; set; }
-        public int ExpiresIn { get; set; }
+        public long ExpiresIn { get; set; }
     }
 }
